[phf[gb;gfb

+++ task3.py
@@ -33,7 +33,7 @@ Enter a number: 10000
 """
 
 x = input("Enter number")
-x = int(x)
+x = float(x)
 if x > 1000 or x ==1000:
     print("The number is or is larger than 1000")
 else:
